Allow for redirection if query string is present
Currently, this extension does not redirect if a query string is present.  If a link is clicked from their email newsletter, which contains utm_* campaign information, this extension will not redirect.

The changes proposed here extend the URL patterns that this extension applies, such that a URL with a query string is accepted.  Furthermore, a check is now performed that evaluates the presence of a question mark (indicating the beginning of the query string), and either adds "?ALLSTEPS" or "&ALLSTEPS" depending on if a query string is already present.